refactor: move signature-related code from account.go to dedicated signature.go file
